+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 2.2.0 (unreleased)
 
 ### New features:
+ * `ERC20`: added internal `_approve(address owner, address spender, uint256 value)`, allowing derived contracts to set the allowance of arbitrary accounts.
 
-### Improvements::
+### Improvements:
  * Upgraded the minimum compiler version to v0.5.2: this removes many Solidity warnings that were false positives.
  * `Counter`'s API has been improved, and is now used by `ERC721` (though it is still in `drafts`).
  * `ERC721`'s transfers are now more gas efficient due to removal of unnecessary `SafeMath` calls.
@@ -13,6 +14,9 @@
 ### Bugfixes:
 
 ### Breaking changes:
+
+## 2.1.2 (2019-17-01)
+ * Removed most of the test suite from the npm package, except `PublicRole.behavior.js`, which may be useful to users testing their own `Roles`.
 
 ## 2.1.1 (2019-04-01)
  * Version bump to avoid conflict in the npm registry.